roles and permissions changes

+++ app/Http/Controllers/Admin/BlogCategoryController.php
@@ -19,55 +19,63 @@ class BlogCategoryController extends Controller
 
     public function index()
     {
-//        if (Auth::user()->can('viewblogcatg')) {
+        if (Auth::user()->can('r-blogcatg'))
+        {
             $categs = BlogCategory::all();
             return view('admin.blogcategs.index', compact('categs'));
-//        } else {
-//            Log::newLog("Usuário tentou acesso area restrita BLOGCATEG, user: " . Auth::user()->name);
-//            return view('admin.layout.403');
-//        }
+
+        } else {
+           Log::newLog("Usuário tentou acesso: R-BLOGCATEG, user: " . Auth::user()->name);
+           return view('admin.layout.403');
+        }
     }
 
     public function store(BlogCateg $request)
     {
-//        if (Auth::user()->can('createblogcatg')) {
+        if (Auth::user()->can('c-blogcatg'))
+        {
             $categ = new BlogCategory();
             $categ->name = $request->get('name');
             $categ->save();
 
-            Log::newLog("Usuário criou novo blog categoria, user: " . Auth::user()->name);
+            Log::newLog("Usuário criou nova categoria, user: " . Auth::user()->name);
             return redirect()->back();
-//        } else {
-//            Log::newLog("Usuário tentou criar area restrita BLOGCATEG, user: " . Auth::user()->name);
-//            return view('admin.layout.403');
-//        }
+
+       } else {
+           Log::newLog("Usuário tentou criar: C-BLOGCATEG, user: " . Auth::user()->name);
+           return view('admin.layout.403');
+       }
 
     }
 
     public function update(Request $request, $id)
     {
-//        if (Auth::user()->can('editblogcatg')) {
+        if (Auth::user()->can('u-blogcatg'))
+        {
             $categ = BlogCategory::findOrFail($id);
             $categ->name = $request->get('name');
             $categ->update();
 
-            Log::newLog("Usuário alterou blog cateria-$id, user: " . Auth::user()->name);
+            Log::newLog("Usuário alterou cateria-{$id}, user: " . Auth::user()->name);
             return redirect()->back();
-//        } else {
-//            Log::newLog("Usuário tentou atualizar area restrita BLOGCATEG-$id, user: " . Auth::user()->name);
-//            return view('admin.layout.403');
-//        }
+
+        } else {
+           Log::newLog("Usuário tentou atualizar: U-BLOGCATEG-{$id}, user: " . Auth::user()->name);
+           return view('admin.layout.403');
+        }
     }
 
     public function destroy($id)
     {
-//        if (Auth::user()->can('deleteblogcatg')) {
+        if (Auth::user()->can('d-blogcatg'))
+        {
             BlogCategory::findOrFail($id)->destroy($id);
-            Log::newLog("Usuário deletou blog cateria-$id, user: " . Auth::user()->name);
+            Log::newLog("Usuário deletou categoria-{$id}, user: " . Auth::user()->name);
             return redirect()->back();
-//        } else {
-//            Log::newLog("Usuário tentou deletar area restrita BLOGCATEG-$id, user: " . Auth::user()->name);
-//            return view('admin.layout.403');
-//        }
+
+        } else {
+            Log::newLog("Usuário tentou deletar: D-BLOGCATEG-{$id}, user: " . Auth::user()->name);
+            return view('admin.layout.403');
+        }
     }
 }
