+++ app/Http/Controllers/Admin/RoleController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\Models\Permission;
 use App\Models\Role;
+use App\Models\Log;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\DB;
@@ -19,29 +20,33 @@ class RoleController extends Controller
 
     public function index()
     {
-//        if (Auth::user()->can('viewperms')) {
+        if (Auth::user()->can('r-roles'))
+        {
             $roles = Role::all();
             $perms = Permission::all();
             return view('admin.roles.index', compact('roles', 'perms'));
-//        } else {
-//            return view('admin.layout.403');
-//        }
+        } else {
+            Log::newLog("Usuário tentou acessar: R-ROLES, user: " . Auth::user()->name);
+            return view('admin.layout.403');
+        }
     }
 
     public function create()
     {
-//        if (Auth::user()->can('createperms')) {
+        if (Auth::user()->can('c-roles'))
+        {
             $perms = Permission::all();
-//            $rolePerm = DB::table('permission_role')->where('permission_role.role_id')->pluck('permission_role.permission_id', 'permission_role.permission_id')->all();
             return view('admin.roles.add', compact('perms', 'rolePerm'));
-//        } else {
-//            return view('admin.layout.403');
-//        }
+        } else {
+            Log::newLog("Usuário tentou criar: C-ROLES, user: " . Auth::user()->name);
+            return view('admin.layout.403');
+        }
     }
 
     public function store(Request $request)
     {
-//        if (Auth::user()->can('createperms')) {
+        if (Auth::user()->can('c-roles'))
+        {
             $role = new Role();
             $role->name = $request->get('name');
             $role->description = $request->get('description');
@@ -53,26 +58,30 @@ class RoleController extends Controller
 
             flash("Regra criada com sucesso.")->success();
             return redirect('dashboard/roles');
-//        } else {
-//            return view('admin.layout.403');
-//        }
+        } else {
+            Log::newLog("Usuário tentou cadastrar: C-ROLES, user: " . Auth::user()->name);
+            return view('admin.layout.403');
+        }
     }
 
     public function edit($id)
     {
-//        if (Auth::user()->can('editperms')) {
+        if (Auth::user()->can('u-roles'))
+        {
             $role = Role::findOrFail($id);
             $perms = Permission::all();
             $rolePerm = DB::table('permission_role')->where('permission_role.role_id', $id)->pluck('permission_role.permission_id', 'permission_role.permission_id')->all();
             return view('admin.roles.edit', compact('role', 'perms', 'rolePerm'));
-//        } else {
-//            return view('admin.layout.403');
-//        }
+        } else {
+            Log::newLog("Usuário tentou editar: U-ROLES-{$id}, user: " . Auth::user()->name);
+            return view('admin.layout.403');
+        }
     }
 
     public function update(Request $request, $id)
     {
-//        if (Auth::user()->can('editperms')) {
+        if (Auth::user()->can('u-roles'))
+        {
             $role = Role::findOrFail($id);
             $role->name = $request->get('name');
             $role->description = $request->get('description');
@@ -85,18 +94,21 @@ class RoleController extends Controller
             }
 
             return redirect('dashboard/roles');
-//        } else {
-//            return view('admin.layout.403');
-//        }
+        } else {
+            Log::newLog("Usuário tentou atualizar: U-ROLES-{$id}, user: " . Auth::user()->name);
+            return view('admin.layout.403');
+        }
     }
 
     public function destroy($id)
     {
-//        if (Auth::user()->can('deleteperms')) {
+        if (Auth::user()->can('d-roles'))
+        {
             DB::table('roles')->where('id', $id)->delete($id);
             return redirect()->back();
-//        } else {
-//            return view('admin.layout.403');
-//        }
+        } else {
+            Log::newLog("Usuário tentou deletar: D-ROLES-{$id}, user: " . Auth::user()->name);
+            return view('admin.layout.403');
+        }
     }
 }
